# public/api/openapi.yaml
openapi: 3.0.0
info:
  title: Book Review API
  description: API for managing books, reviews, and user interactions
  version: 1.0.0
  contact:
    email: y.kabir1@edu.salford.ac.uk
servers:
  - url: /api
    description: API base URL
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        profilePicture:
          type: string
          nullable: true
          example: "/uploads/profile_pictures/john-doe.jpg"
        roles:
          type: array
          items:
            type: string
          example: ["ROLE_USER"]
        banned:
          type: boolean
          example: false
        readingList:
          type: array
          items:
            $ref: '#/components/schemas/Book'
          example: []
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
          example: []
    Book:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        googleBookId:
          type: string
          example: "zyTCAlFPjgYC"
        title:
          type: string
          example: "The Great Gatsby"
        authors:
          type: array
          items:
            type: string
          example: ["F. Scott Fitzgerald"]
        thumbnail:
          type: string
          nullable: true
          example: "https://books.google.com/books/content?id=zyTCAlFPjgYC&printsec=frontcover&img=1&zoom=1&source=gbs_api"
        description:
          type: string
          nullable: true
          example: "The story of the mysteriously wealthy Jay Gatsby and his love for the beautiful Daisy Buchanan."
        pageCount:
          type: integer
          nullable: true
          example: 180
        categories:
          type: array
          items:
            type: string
          example: ["Fiction", "Classics", "Literary"]
        averageRating:
          type: number
          format: float
          nullable: true
          example: 4.5
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
          example: []
    Review:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        content:
          type: string
          example: "A masterpiece of American literature. The prose is beautiful and the story is timeless."
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        createdAt:
          type: string
          format: date-time
          example: "2023-05-15T14:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-05-15T14:30:00Z"
        user:
          $ref: '#/components/schemas/User'
          example: {"id": 1, "name": "John Doe"}
        book:
          $ref: '#/components/schemas/Book'
          example: {"id": 1, "title": "The Great Gatsby"}
        likesCount:
          type: integer
          example: 12
        likedByCurrentUser:
          type: boolean
          example: true
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Validation failed"
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example: {"email": ["This value is not a valid email address."]}
        code:
          type: integer
          example: 400
      required:
        - success
        - message
        - code
  responses:
    UnauthorizedError:
      description: Unauthorized - User is not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Unauthorized"
            code: 401
            errors: {}
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Resource not found"
            code: 404
            errors: {}
...
paths:
...
  /admin/users:
    get:
      summary: List all users
      description: |
        List all users with pagination. Accessible by admins and moderators.
        - Admins and moderators can view all users
        - Pagination parameters are optional
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                        description: Total number of users
                      page:
                        type: integer
                        description: Current page number
                      limit:
                        type: integer
                        description: Number of items per page
                      total_pages:
                        type: integer
                        description: Total number of pages
                      has_next:
                        type: boolean
                        description: Whether there is a next page
                      has_prev:
                        type: boolean
                        description: Whether there is a previous page
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: Access denied. Admin or moderator role required.
                code: 403
                errors: {}
  /admin/users/{id}:
    get:
      summary: Get user details
      description: |
        Get detailed information about a specific user. Accessible by admins and moderators.
        - Admins and moderators can view any user's details
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: Access denied. Admin or moderator role required.
                code: 403
                errors: {}
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      summary: Update user
      description: |
        Update user information. Admin only operation.
        - Only admins can update user information
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: User's name
                email:
                  type: string
                  format: email
                  description: User's email
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: Access denied. Admin role required.
                code: 403
                errors: {}
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      summary: Delete user
      description: |
        Delete a user. Admin only operation.
        - Only admins can delete users
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: Access denied. Admin role required.
                code: 403
                errors: {}
        '404':
          $ref: '#/components/responses/NotFoundError'
  /admin/users/{id}/promote:
    post:
      summary: Promote user to admin
      description: |
        Promote a user to admin role. Admin only operation.
        - Only admins can promote users
        - Cannot promote a user who is already an admin
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User promoted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: Access denied. Admin role required.
                code: 403
                errors: {}
        '404':
          $ref: '#/components/responses/NotFoundError'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: User is already an admin
                code: 400
                errors: {}
  /admin/users/{id}/demote:
    post:
      summary: Demote admin to user
      description: |
        Demote an admin to regular user. Admin only operation.
        - Only admins can demote other admins
        - Cannot demote a user who is not an admin
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User demoted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: Access denied. Admin role required.
                code: 403
                errors: {}
        '404':
          $ref: '#/components/responses/NotFoundError'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: User is not an admin
                code: 400
                errors: {}
  /admin/users/{id}/ban:
    post:
      summary: Ban user
      description: |
        Ban a user. Accessible by admins and moderators.
        - Admins and moderators can ban users
        - Cannot ban a user who is already banned
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User banned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: Access denied. Admin or moderator role required.
                code: 403
                errors: {}
        '404':
          $ref: '#/components/responses/NotFoundError'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: User is already banned
                code: 400
                errors: {}
  /admin/users/{id}/unban:
    post:
      summary: Unban user
      description: |
        Unban a user. Accessible by admins and moderators.
        - Admins and moderators can unban users
        - Cannot unban a user who is not banned
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User unbanned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: Access denied. Admin or moderator role required.
                code: 403
                errors: {}
        '404':
          $ref: '#/components/responses/NotFoundError'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: User is not banned
                code: 400
                errors: {}
...